third_party/blink/web_tests/fast/hidpi/static/mousewheel-scroll-amount.html

<!DOCTYPE html>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../virtual/percent-based-scrolling/resources/percent-based-util.js"></script>

<style type="text/css">
    body {
        height: 2000px;
        width: 2000px;
    }
</style>

<script>
async function reset() {
    document.scrollingElement.scrollTop = 0;
    document.scrollingElement.scrollLeft = 0;
    await waitForCompositorCommit();
}

var position = {x: 20, y: 20};
promise_test(async () => {
    await reset();
    await wheelTick(0, 1, position);
    await waitForAnimationEndTimeBased( () => {
        return document.scrollingElement.scrollTop;
    });
    const {x: expectedScrollX, y: expectedScrollY} = 
        calculateExpectedScroll(document.scrollingElement, 0, pixelsPerTick());
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, expectedScrollX, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, expectedScrollY, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), expectedScrollX);
        assert_equals(Math.round(document.scrollingElement.scrollTop), expectedScrollY);
    }
}, "One vertical wheel tick scrolls the right number of pixels in high dpi mode.");

promise_test(async () => {
    await reset();
    await wheelTick(1, 0, position);
    await waitForAnimationEndTimeBased( () => {
        return document.scrollingElement.scrollLeft;
    });
    const {x: expectedScrollX, y: expectedScrollY} = 
        calculateExpectedScroll(document.scrollingElement, pixelsPerTick(), 0);
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, expectedScrollX, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, expectedScrollY, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), expectedScrollX);
        assert_equals(Math.round(document.scrollingElement.scrollTop), expectedScrollY);
    }
}, "One horizontal wheel tick scrolls the right number of pixels in high dpi mode.");

promise_test(async () => {
    await reset();
    await wheelTick(0, 2, position);
    await waitForAnimationEndTimeBased( () => {
        return document.scrollingElement.scrollTop;
    });
    const {x: expectedScrollX, y: expectedScrollY} = 
        calculateExpectedScroll(document.scrollingElement, 0, pixelsPerTick() * 2);
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, expectedScrollX, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, expectedScrollY, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), expectedScrollX);
        assert_equals(Math.round(document.scrollingElement.scrollTop), expectedScrollY);
    }
}, "Two vertical wheel ticks scroll the right number of pixels in high dpi mode.");

promise_test(async () => {
    await reset();
    await wheelTick(2, 0, position);
    await waitForAnimationEndTimeBased( () => {
        return document.scrollingElement.scrollLeft;
    });
    const {x: expectedScrollX, y: expectedScrollY} = 
        calculateExpectedScroll(document.scrollingElement, pixelsPerTick() * 2, 0);
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, expectedScrollX, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, expectedScrollY, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), expectedScrollX);
        assert_equals(Math.round(document.scrollingElement.scrollTop), expectedScrollY);
    }
}, "Two horizontal wheel ticks scroll the right number of pixels in high dpi mode.");
</script>
